add support for human36 model (still work in progress)
At this point, the anatomical lengths, tags and inertia data have been
added and the 2 former have been tested by checking against the HuMAnS
toolbox results. However the inertial data are not used yet, and only a
subpart of the kinematic model written. The KinematicData.maple file was
added for reference, it should be deleted when the conversion is
finished.