introducing MovingSubframe class
SubFrames come in two flavors: the constant and the moving ones.

Previously a single class (SubFrame) fitted both needs. When the
SubFrame needed to be moved, its private _bpose property was modified
(from outside the class).

This commit introduces a formal with two classes : MovingSubFrame (for
the moving ones) and SubFrame (the constant ones).

The moving ones are only used in PointContact class, to update the
contact position within the each body frame. They might also be
helpful to users who need (moving) target frames etc.

Everywhere else the subframes are constant.

This commit leads to cleaner code (no more exterior private acces)
and gives a clean way to know which subframes should be saved (as
trajectories) for a later visualization.