removing ObservableWorld
The ObservableWorld was meant to let plugins (observers) capture events
from world creation (register, add_link) and simulation (init, update).

In particular, the visu_osg.Drawer used to create the scene graph
corresponding to the world during the world creation. It was then
possible to view the world while it was built.

However, this version of the Drawer is no more, and the world creation
method was extended (with replace_joint, name_frames), which would make
the task of rewritting it hardier.

Thus, this commit removes the ObservableWorld class. The observers
should now be given as a third argument to the simulate function.

This commit also

- makes the WorldObserver.finish() method  mandatory.
- adds the world and timeline to the WorldObserver.init() method