updated install doc to target ubuntu 10.4
Ubuntu 10.4 (lucid lynx) ships with a package for h5py, we don't need to
build it anymore.

As the lucid lynx is a LTS release and karmic koala was not, it's not
likely that many users will stay on karmic. Then, I dropped the installation
instructions for karmic, to avoid clutter.